+++ modules/normalizeRoutes.js
@@ -1,24 +1,30 @@
 import React from 'react';
-import { map } from 'lodash';
+import { map, partial } from 'lodash';
+import { getNestedState } from './nestedState';
+import { ROOT_DEPTH } from './constants';
 
 const defaultLoader = () => <div/>;
 
-export default (routes, force = false) => {
-  return map(routes, route => {
+export default (routes, props, startIndex = ROOT_DEPTH, force = false) => {
+  return map(routes, (route, index) => {
     if (!route.normalized || force) {
-      route.normalized = true;
-
-      if (!route.reducer) {
-        route.reducer = state => state;
+      if (props && route.onLeave) {
+        const { location: { query }, params, store } = props;
+        const getReducerState = () => getNestedState(store.getState(), startIndex + index);
+        route.onLeave = partial(route.onLeave, {
+          getReducerState,
+          params,
+          query,
+        });
       }
 
-      if (!route.loader) {
-        route.loader = defaultLoader;
-      }
+      if (!route.reducer) route.reducer = state => state;
+      if (!route.loader) route.loader = defaultLoader;
 
-      route.blockRender = true;
       route.loading = true;
       route.loadingError = null;
+      route.blockRender = true;
+      route.normalized = true;
     }
 
     return route;