+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 - [x] Handle deserialize in constructor not loadStateOnClient.
 - [x] Rename reducerData to getReducerState
-- [ ] Change asyncLeave to onLeave & call with extra params.
+- [x] Change asyncLeave to onLeave & call with extra params.
 - [ ] Use state for injected params.
 - [ ] Route->Route+queryParams trigger asyncEnter (error / redirect).